sticky scrollers 95% done. 'RedTrail' added.
I've finalized the sticky scrollers to about 95%, still need to tweak sizing to allow for different sized content and see if I can hide content that leaves the top of the page.  RedTrail is a very quick mouse over project I created, but I've reduced the number of divs from 2500 to 400 as performance was suffering with other elements on screen.  I've also changed the functionality of it to reveal an image underneath, and will further tweak the transitions and possibly div sizes.  Image is just placeholder.