changed redtrail to an image
I need to find more dummy images, it's suuuuper confusing to look at right now!  The fixed aspect of the redtrail makes scrolling super weird.  Need to take it out of the sticky and get another sticky to cover it up instead - the interactive divs scroll but the image stays in place, but image that is revealed DOES scroll.  Also need to cover up the area above and below the stickies to make it more coherent if I'm not going to have it as a full-height element.