Missed a commit or 2... A lot has happened!
Every time I think I've finished the stickies, I try to add content and break things.  This time I've REALLY finished them.  They're scalable and positional (but breakpoints will definitely be needed) and I've done a full write-up in the css file on what's going on.  I guess that's what a README is for?  Whoops.  There are two commits I thought I'd done, but they're not in the log, hrmmm.  I've started the very basics of adding content, have added in a colour pallete and self-referential 60/30/10 gradient.  I'm quite happy with it while it's on my laptop screen, but it's QUITE a different colour on my big monitor (very purple).  From here on out I REALLY need to just focus on content and no more functional stuff.  If I get done with content THEN I can tinker, but not before!